refactor: move listener types to types directory
